Fixes missing s:window and aqua declarations
Adds missing declarations for s:window in the Tomorrow, Tomorrow Night
Bright, Tomorrow Night Blue, and Tomorrow Night Eighites themes. For the
Tomorrow Night Bright and Tomorrow Night Eighties themes, the s:window
value was set to the same value as was set in Tomorrow Night. For the
Tomorrow Night Blue and Tomorrow themes, the current line color as
specified at https://github.com/ChrisKempson/Tomorrow-Theme was used

Also added a missing aqua declaration to the Tomorrow theme. This
was set to the value specified in the above link.